feat: allow user configuration
With this change we are finally allowing some customization of options via configuration file. This
first iteration makes so the journal file path can be set, and if not, uses the user home as base

BREAKING CHANGE: as we are now using safe defaults, previous users will
need to move their current journal file to the desired path, or setup
the configuration options in the config.yaml file.

docs(readme.md): add configuration section